refactor: move compareVersions to utils
